Display some additional status information in the Goals panel
When a proof is in progress, add a bit to the bottom of the goals panel
which shows the name of the active proof, the names of any incomplete
proofs further up the lemma stack (if any), and the current
module/section. (This is what the "Status" xml command exposes.)

When no proof is in progress, show "No proof in progress" (and the
current module/section) instead of an empty window.